+++ Specs/DynamicScene/DynamicPointVisualizerSpec.js
@@ -52,8 +52,8 @@ defineSuite([
         visualizer = new DynamicPointVisualizer(scene, dynamicObjectCollection);
         expect(visualizer.getScene()).toEqual(scene);
         expect(visualizer.getDynamicObjectCollection()).toEqual(dynamicObjectCollection);
-        expect(scene.getPrimitives().getLength()).toEqual(1);
-        var billboardCollection = scene.getPrimitives().get(0);
+        expect(scene.primitives.length).toEqual(1);
+        var billboardCollection = scene.primitives.get(0);
         expect(billboardCollection instanceof BillboardCollection).toEqual(true);
     });
 
@@ -85,7 +85,7 @@ defineSuite([
         var testObject = dynamicObjectCollection.getOrCreateObject('test');
         testObject.position = new ConstantProperty(new Cartesian3(1234, 5678, 9101112));
         visualizer.update(new JulianDate());
-        var billboardCollection = scene.getPrimitives().get(0);
+        var billboardCollection = scene.primitives.get(0);
         expect(billboardCollection.getLength()).toEqual(0);
     });
 
@@ -98,7 +98,7 @@ defineSuite([
         point.show = new ConstantProperty(true);
 
         visualizer.update(new JulianDate());
-        var billboardCollection = scene.getPrimitives().get(0);
+        var billboardCollection = scene.primitives.get(0);
         expect(billboardCollection.getLength()).toEqual(0);
     });
 
@@ -108,7 +108,7 @@ defineSuite([
         var dynamicObjectCollection = new DynamicObjectCollection();
         visualizer = new DynamicPointVisualizer(scene, dynamicObjectCollection);
 
-        var billboardCollection = scene.getPrimitives().get(0);
+        var billboardCollection = scene.primitives.get(0);
         expect(billboardCollection.getLength()).toEqual(0);
 
         var testObject = dynamicObjectCollection.getOrCreateObject('test');
@@ -163,7 +163,7 @@ defineSuite([
     it('clear hides billboards.', function() {
         var dynamicObjectCollection = new DynamicObjectCollection();
         visualizer = new DynamicPointVisualizer(scene, dynamicObjectCollection);
-        var billboardCollection = scene.getPrimitives().get(0);
+        var billboardCollection = scene.primitives.get(0);
         expect(billboardCollection.getLength()).toEqual(0);
         var testObject = dynamicObjectCollection.getOrCreateObject('test');
         var time = new JulianDate();
@@ -187,7 +187,7 @@ defineSuite([
         var dynamicObjectCollection = new DynamicObjectCollection();
         visualizer = new DynamicPointVisualizer(scene, dynamicObjectCollection);
 
-        var billboardCollection = scene.getPrimitives().get(0);
+        var billboardCollection = scene.primitives.get(0);
         expect(billboardCollection.getLength()).toEqual(0);
 
         var testObject = dynamicObjectCollection.getOrCreateObject('test');
@@ -220,7 +220,7 @@ defineSuite([
         visualizer = new DynamicPointVisualizer(scene, dynamicObjectCollection);
 
         var time = new JulianDate();
-        var billboardCollection = scene.getPrimitives().get(0);
+        var billboardCollection = scene.primitives.get(0);
 
         visualizer.update(time);
         expect(billboardCollection.getLength()).toEqual(1);